Calculate data to construct bivariate partial dependence plots

Calculate data to construct bivariate partial dependence for two predictor set

Usage

data_abund_bpdp(
  model,
  predictors,
  resolution = 50,
  training_data = NULL,
  invert_transform = NULL,
  response_name = NULL,
  training_boundaries = NULL,
  projection_data = NULL
)

Arguments

model

object returned by any fit_abund or tune_abund family functions

predictors

character. Vector with two predictor name(s) to plot. If NULL all predictors will be plotted. Default NULL

resolution

numeric. Number of equally spaced points at which to predict continuous predictors. Default 50

training_data

data.frame or tibble. Database with response (0,1) and predictor values used to fit a model. Default NULL

invert_transform

logical. Invert transformation of response variable. Useful for those cases that the response variable was transformed with one of the method in adm_transform. Default NULL

response_name

character. Name of the response variable. Default NULL

training_boundaries

character. Plot training conditions boundaries based on training data (i.e., presences, presences and absences, etc). If training_boundaries = "convexh", function will delimit training environmental region based on a convex-hull. If training_boundaries = "rectangle", function will delimit training environmental region based on four straight lines. If used any methods it is necessary provide data in training_data argument. If NULL all predictors will be used. Default NULL.

projection_data

SpatRaster. Raster layer with environmental variables used for model projection. Default NULL

Value

A list with two tibbles "pdpdata" and "resid".

pspdata: has data to construct partial dependence surface plot, the first two column includes values of the selected environmental variables, the third column the predicted suitability.

training_boundaries: has data to plot boundaries of training data.

See also