Fit and validate Convolutional Neural Network Model

This function is used to fit a convolutional neural network (CNN) model for abundance.

Usage

fit_abund_cnn(
  data,
  response,
  predictors,
  predictors_f = NULL,
  x,
  y,
  rasters,
  sample_size,
  partition,
  predict_part = FALSE,
  learning_rate = 0.01,
  n_epochs = 10,
  batch_size = 32,
  validation_patience = 2,
  fitting_patience = 5,
  custom_architecture = NULL,
  verbose = TRUE
)

Arguments

data

tibble or data.frame. Database with response, predictors, and partition values

response

character. Column name with species abundance.

predictors

character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("temp", "precipt", "sand")

predictors_f

character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")

x

character. The name of the column containing longitude information for each observation.

y

character. The name of the column containing latitude information for each observation.

rasters

a terra SpatRaster object. A raster containing the predictor variables to be cropped around each observation.

sample_size

numeric. The dimension, in pixels, of raster samples. See cnn_make_samples beforehand. Default c(11,11)

partition

character. Column name with training and validation partition groups.

predict_part

logical. Save predicted abundance for testing data. Default = FALSE

learning_rate

numeric. The size of the step taken during the optimization process. Default = 0.01

n_epochs

numeric. How many times the learning algorithm will work through the training set. Default = 10

batch_size

numeric. A batch is a subset of the training set used in a single iteration of the training process. The size of each batch is referred to as the batch size. Default = 32

validation_patience

fitting_patience

custom_architecture

a Torch nn_module_generator object. A neural network architecture to be used instead of the internal default one. Default NULL

verbose

logical. If FALSE, disables all console messages. Default TRUE

Value

A list object with:

model: A "luz_module_fitted" object from luz (torch framework). This object can be used to predicting.

predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.

performance: Averaged performance metrics (see adm_eval).

performance_part: Performance metrics for each replica and partition.

predicted_part: Observed and predicted abundance for each test partition.