Generate architectures for Deep Neural Network

Generate architectures for Deep Neural Network

Usage

generate_dnn_architecture(
  number_of_features = 7,
  number_of_outputs = 1,
  number_of_hidden_layers = 2,
  hidden_layers_size = c(14, 7),
  batch_norm = TRUE,
  dropout = 0.2,
  verbose = FALSE
)

Arguments

number_of_features

numeric. Value that specifies the number of features in the dataset.

number_of_outputs

numeric. Value that specifies the number of outputs.

number_of_hidden_layers

numeric. Number of hidden layers in the neural network. Default 2.

hidden_layers_size

numeric vector. Size of each hidden layer in the neural network. Default c(14, 7).

batch_norm

logical. Whether to include batch normalization layers. Default TRUE.

dropout

logical. Specifies whether dropout is included in the architecture. Default FALSE.

verbose

logical. Whether to print the architecture. Default FALSE.

Value

A list containing:

net: a list of generated architectures.

arch: a list of architecture string

See also